pyproject.toml
[tool.poetry]
name = "slackbot-gitlab-integration"
version = "0.1.0"
description = ""
authors = ["sommit <sommit@rentspree.com>"]
readme = "README.md"
packages = [{include = "mr_scapper"}]

[tool.poetry.dependencies]
python = "^3.10"
slackclient = "^2.9.4"
python-gitlab = "^3.10.0"
timeago = "^1.0.16"
python-dateutil = "^2.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
